+++ packages/theme/src/cli/services/pull.ts
@@ -1,7 +1,8 @@
 import {downloadTheme} from '../utilities/theme-downloader.js'
 import {hasRequiredThemeDirectories, mountThemeFileSystem} from '../utilities/theme-fs.js'
 import {currentDirectoryConfirmed, themeComponent} from '../utilities/theme-ui.js'
-import {Checksum, Theme} from '@shopify/cli-kit/node/themes/types'
+import {rejectLiquidChecksums} from '../utilities/asset-checksum.js'
+import {Theme} from '@shopify/cli-kit/node/themes/types'
 import {AdminSession} from '@shopify/cli-kit/node/session'
 import {fetchChecksums} from '@shopify/cli-kit/node/themes/api'
 import {renderSuccess} from '@shopify/cli-kit/node/ui'
@@ -62,33 +63,6 @@ export async function pull(theme: Theme, session: AdminSession, options: PullOpt
         },
       ],
     ],
-  })
-}
-
-/**
- * Filters out generated asset files from a list of theme checksums.
- *
- * The checksums API returns entries for both original and generated files. For
- * instance, if there's a Liquid file 'assets/basic.css.liquid', the API will
- * return entries for both 'assets/basic.css.liquid' and the generated
- * 'assets/basic.css' with the same checksum.
- *
- * Example:
- *   - key: 'assets/basic.css',        checksum: 'e4b6aac5f2af8ea6e197cc06102186e9'
- *   - key: 'assets/basic.css.liquid', checksum: 'e4b6aac5f2af8ea6e197cc06102186e9'
- *
- * This function filters out the generated files (like 'assets/basic.css'),
- * as these are not needed for theme comparison.
- */
-export function rejectLiquidChecksums(themeChecksums: Checksum[]) {
-  return themeChecksums.filter(({key}) => {
-    const isStaticAsset = key.startsWith('assets/')
-
-    if (isStaticAsset) {
-      return !themeChecksums.some((checksum) => checksum.key === `${key}.liquid`)
-    }
-
-    return true
   })
 }
