+++ packages/theme/src/cli/utilities/asset-checksum.ts
@@ -1,4 +1,5 @@
 import {isThemeAsset, isJson, readThemeFile, isTextFile} from './theme-fs.js'
+import {Checksum} from '@shopify/cli-kit/node/themes/types'
 import {fileHash} from '@shopify/cli-kit/node/crypto'
 
 export async function checksum(root: string, path: string) {
@@ -52,4 +53,31 @@ function md5(content: string) {
   const buffer = Buffer.from(content)
 
   return fileHash(buffer)
+}
+
+/**
+ * Filters out generated asset files from a list of theme checksums.
+ *
+ * The checksums API returns entries for both original and generated files. For
+ * instance, if there's a Liquid file 'assets/basic.css.liquid', the API will
+ * return entries for both 'assets/basic.css.liquid' and the generated
+ * 'assets/basic.css' with the same checksum.
+ *
+ * Example:
+ *   - key: 'assets/basic.css',        checksum: 'e4b6aac5f2af8ea6e197cc06102186e9'
+ *   - key: 'assets/basic.css.liquid', checksum: 'e4b6aac5f2af8ea6e197cc06102186e9'
+ *
+ * This function filters out the generated files (like 'assets/basic.css'),
+ * as these are not needed for theme comparison.
+ */
+export function rejectLiquidChecksums(themeChecksums: Checksum[]) {
+  return themeChecksums.filter(({key}) => {
+    const isStaticAsset = key.startsWith('assets/')
+
+    if (isStaticAsset) {
+      return !themeChecksums.some((checksum) => checksum.key === `${key}.liquid`)
+    }
+
+    return true
+  })
 }
